Extract superclass: CommandLineTool
